Add catch for url being null in Fanart Generator Fixes #61

+++ fanartGenerator.js
@@ -2,13 +2,13 @@ var images = require("images");
 var logger = require("./logger");
 
 function generateFanart(imdb, resolution, callback) {
-  logger.Debug("=== Generating Fanart ===");
-  var width = 1920;
-  var height = 1080;
-  if (resolution == '720'){
-    width = 1280;
-    height = 720;
-  }
+    logger.Debug("=== Generating Fanart ===");
+    var width = 1920;
+    var height = 1080;
+    if (resolution == '720') {
+        width = 1280;
+        height = 720;
+    }
     var fs = require('fs');
     if (fs.existsSync('assets/cache/' + imdb + '-' + resolution + '.jpg')) {
         logger.Debug('Fanart already downloaded, serving.');
@@ -16,46 +16,53 @@ function generateFanart(imdb, resolution, callback) {
     } else {
         var API = require('./MoviesAPI');
         var movies = API.getFanart(imdb, function(url) {
-            var http = require('https');
-            var file = fs.createWriteStream("tmp.jpg");
-            var request = http.get(url, function(response) {
-              response.pipe(file);
-              file.on('finish', function() {
-                images('tmp.jpg').resize(width, height).draw(images('assets/thumbnails/gradient_' + resolution + '.png'), 0, 0).save('assets/cache/' + imdb + '-' + resolution + '.jpg');
-                callback('cache/' + imdb + '-' + resolution + '.jpg');
-                fs.unlink('tmp.jpg');
-              });
-            });
+            if (url == null) {
+                callback('thumbnails/Background_blank_' + resolution + '.jpg');
+            } else {
+                var http = require('https');
+                var file = fs.createWriteStream("tmp.jpg");
+                var request = http.get(url, function(error, response, body) {
+                    if (!error && response.statusCode === 200) {
+                        response.pipe(file);
+                        file.on('finish', function() {
+                            images('tmp.jpg').resize(width, height).draw(images('assets/thumbnails/gradient_' + resolution + '.png'), 0, 0).save('assets/cache/' + imdb + '-' + resolution + '.jpg');
+                            callback('cache/' + imdb + '-' + resolution + '.jpg');
+                            fs.unlink('tmp.jpg');
+                        });
+                    }
+                });
+            }
         });
     }
 }
+
 function generateFanartTV(url, imdb, resolution, callback) {
     logger.Debug("=== Generating Fanart for TV ===");
     var width = 1920;
     var height = 1080;
-    if (resolution == '720'){
-      width = 1280;
-      height = 720;
+    if (resolution == '720') {
+        width = 1280;
+        height = 720;
     }
     var fs = require('fs');
     if (fs.existsSync('assets/cache/' + imdb + '-' + resolution + '.jpg')) {
         logger.Debug('Fanart already downloaded, serving.');
         callback('cache/' + imdb + '-' + resolution + '.jpg');
     } else {
-      var http = require('https');
-      if (url == 'null') {
-          callback('thumbnails/Background_blank_1080.jpg');
-       } else {
-           var file = fs.createWriteStream("tmp.jpg");
-           var request = http.get(url, function(response) {
-              response.pipe(file);
-               file.on('finish', function() {
-                   images('tmp.jpg').resize(width, height).draw(images('assets/thumbnails/gradient_' + resolution + '.png'), 0, 0).save('assets/cache/' + imdb + '-' + resolution + '.jpg');
-                   callback('cache/' + imdb + '-' + resolution + '.jpg');
-                   fs.unlink('tmp.jpg');
-               });
-          });
-       }
+        var http = require('https');
+        if (url == null) {
+            callback('thumbnails/Background_blank_' + resolution + '.jpg');
+        } else {
+            var file = fs.createWriteStream("tmp.jpg");
+            var request = http.get(url, function(response) {
+                response.pipe(file);
+                file.on('finish', function() {
+                    images('tmp.jpg').resize(width, height).draw(images('assets/thumbnails/gradient_' + resolution + '.png'), 0, 0).save('assets/cache/' + imdb + '-' + resolution + '.jpg');
+                    callback('cache/' + imdb + '-' + resolution + '.jpg');
+                    fs.unlink('tmp.jpg');
+                });
+            });
+        }
     }
 }
